# Bursts feature performance improvement