@timonmerk 2024-01-28
That's amazing! You simplified a function quite a lot! It took me some time to understand, (and I changed the name of the `burst_states` to `burst_time_points`) but iterating only every second index of the burst transition is actually quite smart.

@ali-dzaye This adapted function might be interesting for you. The calculation results in the same length/amplitude outputs as our previous function but is calculated much more efficiently.